\name{FourierIntegral-package}
\alias{FourierIntegral-package}
\alias{FourierIntegral}
\docType{package}
\title{
\packageTitle{FourierIntegral}
}
\description{
\packageDescription{FourierIntegral}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FourierIntegral}
\packageIndices{FourierIntegral}
Function \verb|FI| computes the marginal likelihood of any given posterior sample using the Fourier integral method described in Rotiroti & Walker (2022), based on one single R. Function \verb|FIautoR| computes the marginal likelihood based on multiple Rs specefied by users and automatically selects the R based on the median of all computation results or a change-point model.
}
\author{
\packageAuthor{Siyang Li}

Maintainer: \packageMaintainer{Siyang Li}
}
\references{
Rotiroti & Walker (2022), Ho & Walker (2021)
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package, marginal likelihood, Fourier transform, Bayesian inference }
\seealso{
  \code{\link[<density>:<stats>-package]{<density>}}
  \code{\link[<kde2d>:<MASS>-package]{<kde2d>}}
  \code{\link[<bkde>:<KernSmooth>-package]{<bkde>}}
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
